# Conflicts:
#	CHANGELOG.md
#	api/src/main/java/org/allaymc/api/player/Player.java
#	server/src/main/java/org/allaymc/server/AllayServer.java
#	server/src/main/java/org/allaymc/server/player/AllayLoginData.java
#	server/src/main/java/org/allaymc/server/player/AllayPlayer.java
#	server/src/main/java/org/allaymc/server/player/AllayPlayerManager.java

+++ server/src/main/java/org/allaymc/server/world/chunk/ScheduledUpdateInfo.java
@@ -26,7 +26,7 @@ public final class ScheduledUpdateInfo {
     public static ScheduledUpdateInfo fromNBT(NbtMap nbt) {
         return new ScheduledUpdateInfo(
                 new Vector3i(nbt.getInt("x"), nbt.getInt("y"), nbt.getInt("z")),
-                nbt.getInt("time")
+                nbt.getLong("time")
                 // blockState is not used, just ignore it
         );
     }
